% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidynetwork.R
\name{get_tidynetwork}
\alias{get_tidynetwork}
\title{Obtain a tidy sf from OpenStreetMap data}
\usage{
get_tidynetwork(..., simplify_highway = TRUE, highway_filter)
}
\arguments{
\item{...}{parameters passed to \code{osmextract::oe_get_network()}}

\item{simplify_highway}{logical, whether to simplify the highway values by removing the "_link" suffix and filtering by \code{highway_filter}}

\item{highway_filter}{character vector of highway types to keep, if \code{simplify_highway} is TRUE}
}
\value{
a \code{sf} object with standardised highway and oneway values
}
\description{
Obtain a tidy sf from OpenStreetMap data
}
\examples{
\dontrun{
my_area <- sf::st_point(c(-1.6005470549372385,53.836053590512215)) |>
  sf::st_sfc(crs = 4326) |>
 sf::st_buffer(units::set_units(1, "km"))
tidynet_sf <- get_tidynetwork(place = my_area, mode = "driving")
}
}
