% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidynetwork.R
\name{oe_get_sfnetwork}
\alias{oe_get_sfnetwork}
\title{Obtain a sfnetwork object from OpenStreetMap data}
\usage{
oe_get_sfnetwork(
  ...,
  directed = FALSE,
  simplify_highway = TRUE,
  highway_filter = c("motorway", "trunk", "primary", "secondary", "tertiary",
    "unclassified", "residential")
)
}
\arguments{
\item{...}{parameters passed to \code{osmextract::oe_get_network()}}

\item{directed}{logical, whether to return a directed sfnetwork object (default is FALSE)}

\item{simplify_highway}{logical, whether to simplify the highway values by removing the "_link" suffix and filtering by \code{highway_filter}}

\item{highway_filter}{character vector of highway types to keep, if \code{simplify_highway} is TRUE}
}
\value{
a sfnetwork object
}
\description{
Obtain a sfnetwork object from OpenStreetMap data
}
\examples{
\dontrun{
my_area <- sf::st_point(c(-1.6005470549372385,53.836053590512215)) |>
  sf::st_sfc(crs = 4326) |>
  sf::st_buffer(units::set_units(1, "km"))

sfnet_directed <- oe_get_sfnetwork(place = my_area, mode = "driving", directed = TRUE)

sfnet_undirected <- oe_get_sfnetwork(place = my_area, mode = "driving", directed = FALSE)
}
}
