% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidynetwork.R
\name{net_2_sfnet_undirected}
\alias{net_2_sfnet_undirected}
\title{Convert a spatial network to an undirected sfnetwork}
\usage{
net_2_sfnet_undirected(net_sf)
}
\arguments{
\item{net_sf}{a \code{sf} object representing a spatial network}
}
\value{
a \code{sfnetwork} object
}
\description{
Convert a spatial network to an undirected sfnetwork
}
\examples{
\dontrun{
my_area <- sf::st_point(c(-1.6005470549372385,53.836053590512215)) |>
 sf::st_sfc(crs = 4326) |>
sf::st_buffer(units::set_units(1, "km"))
net_sf <- get_tidynetwork(place = my_area, mode = "driving")
sfnet_undirected <- net_2_sfnet_undirected(net_sf)
}

}
