% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidynetwork.R
\name{oe_get_dodgrnetwork}
\alias{oe_get_dodgrnetwork}
\title{Obtain a weighted_streetnet from OpenStreetMap data}
\usage{
oe_get_dodgrnetwork(
  ...,
  highway_filter = c("motorway", "trunk", "primary", "secondary", "tertiary",
    "unclassified", "residential")
)
}
\arguments{
\item{...}{parameters passed to \code{osmextract::oe_get_network()} and \code{dodgr::weight_streetnet()}, excluding \code{x} and \code{id_col} for the latter.}

\item{highway_filter}{string vector of highway types to keep. By default, it includes "motorway", "trunk", "primary", "secondary", "tertiary", "unclassified", and "residential".}
}
\value{
a \code{dodgr_streetnet} object
}
\description{
This function is a wrapper around \code{osmextract::oe_get_network()}
that returns a \verb{dodgr_streetnet`` object. It performs simplification of the highway values, filters by highway types and standardises the }oneway\verb{attribute as well as applying the implied oneway restriction based on the}junction`` tag values.
}
\examples{
\dontrun{
 my_area <- sf::st_point(c(-1.6005470549372385, 53.836053590512215)) |>
   sf::st_sfc(crs = 4326) |>
   sf::st_buffer(units::set_units(1, "km"))
 dodgr_graph <- oe_get_dodgrnetwork(
   place = my_area,
   mode = "driving",
   wt_profile = "motorcar",
   left_side = TRUE
 )
}

}
