% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_oneway}
\alias{tidy_oneway}
\title{Tidy the oneway values in a osm network}
\usage{
tidy_oneway(net_raw, implied_oneway = TRUE)
}
\arguments{
\item{net_raw}{a \code{sf} object representing a spatial network with the \code{oneway} and \code{junction} columns}

\item{implied_oneway}{logical, whether to apply the implied \code{oneway} restriction}
}
\value{
a \code{sf} object with standardised oneway values
}
\description{
This helper function standardises the oneway values in a osm network.
It also applies the implied \code{oneway} tag restriction based on the \code{junction}
tag values if specified.
}
\details{
See \href{https://wiki.openstreetmap.org/wiki/Key:oneway#Implied_oneway_restriction}{oneway tag Wiki}.
}
\examples{
\dontrun{
my_area <- sf::st_point(c(-1.6005470549372385,53.836053590512215)) |>
  sf::st_sfc(crs = 4326) |>
 sf::st_buffer(units::set_units(1, "km"))
sf_net <- oe_get_network(place = my_area, mode = "driving")
sf_net_tidy <- tidy_oneway(sf_net, implied_oneway = TRUE)
}
}
